Added my first project (a basic calculator)
Added my first basic project to the repository. It might seem like a silly thing, but building this very simple calculator was a big deal for me and it's the first time I feel like I've turned a corner with JavaScript as it's one of those languages that I've always had some trouble with compared to something like Python, which for some reason comes really easily to me.