Made a "Counter Program" and a "Dice Roller"
Made a couple of small JavaScript programs - a "Counter Program" and a "Dice Roller". These were inspired by a BroCode video on basic JavaScript skills but I've tried to put my own spin on them, either stylistically or functionally. Will probably continue to follow it along and do the other projects in the video as well, but will aspire to put my own spin on those as well.